Nixos: Let the user override nameserver settings.
When the user specifies the networking.nameservers setting in the
configuration file, it must take precedence over automatically
derived settings.

This patch prevents the service.bind and dnsmasq.resolveLocalQueries
settings from overriding the users' settings.

Also, when the user specifies a domain to search, it must be set in
the resolver configuration, even if the user does not specify any
nameservers.